requirements.txt
streamlit
google-generativeai
openpyxl
python-dotenv
